*changed ctsalvage to int

+++ AdjustedMechSalvage/AdjustedMechSalvage/HolderClasses.cs
@@ -3,7 +3,7 @@
         public bool ownMechsForFree = false;
         public float ejectRecoveryBonus = 0.75f;
         public float incapacitatedRecoveryBonus = 0.5f;
-        public float centerTorsoSalvageValue = 1;
+        public int centerTorsoSalvageValue = 1;
     }
 
     public class Fields {
